+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * stool is no long responsible to manage checkouts - use your standard tools instead
   * dumped `create` and `refresh` command
   * dumped `refresh` stage config
-  * changed `remove` command to leave the project as-is
+  * changed `remove` command to leave the project as-is; the current directory is also left as-is
   * dumped svn credentials handling
   * dumped macros